requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django-cors-headers==3.0.2
keyring==10.6.0
keyrings.alt==3.0
Pillow==6.0.0
Pygments==2.3.1
python-debian==0.1.32
requests==2.22.0
